api/requirements.txt
asgiref==3.7.2
Django==4.2.4
django-ckeditor==6.3.0
django-cors-headers==4.2.0
django-js-asset==2.1.0
djangorestframework==3.14.0
Pillow==10.0.0
pytz==2023.3
sqlparse==0.4.4
tzdata==2023.3
